refactor: rename and move files in sector loader
