+++ viewer/packages/cad-geometry-loaders/index.ts
@@ -4,7 +4,7 @@
 
 export { CadModelSectorBudget } from './src/CadModelSectorBudget';
 
-export { V8SectorRepository } from '../sector-loader/src/CachedRepository';
+export { V8SectorRepository } from '../sector-loader/src/V8SectorRepository';
 
 export { CadLoadingHints } from './src/CadLoadingHints';
