+++ viewer/packages/sector-loader/src/v8/SimpleAndDetailedToSector3D.ts
