+++ viewer/packages/sector-loader/src/v8/groupMeshesByNumber.ts
