+++ viewer/packages/sector-loader/src/v8/sectorUtilities.ts
