+++ viewer/packages/sector-loader/tests/groupMeshesByNumber.test.ts
@@ -2,7 +2,7 @@
  * Copyright 2021 Cognite AS
  */
 
-import { groupMeshesByNumber } from '../src/groupMeshesByNumber';
+import { groupMeshesByNumber } from '../src/v8/groupMeshesByNumber';
 
 describe('groupMeshesByNumber', () => {
   test('empty input yields no result', () => {
