+++ viewer/packages/sector-loader/tests/sectorUtilities.test.ts
@@ -6,7 +6,7 @@ import * as THREE from 'three';
 import { SectorMetadata, SectorGeometry, TriangleMesh } from '@reveal/cad-parsers';
 
 import { createEmptySector } from '../../../test-utilities/src/emptySector';
-import { consumeSectorDetailed, consumeSectorSimple } from '../src/sectorUtilities';
+import { consumeSectorDetailed, consumeSectorSimple } from '../src/v8/sectorUtilities';
 import { SectorQuads, createMaterials, RenderMode } from '@reveal/rendering';
 
 import 'jest-extended';
